Move uart i/o from interrupt into os task
Before this commit reading bytes from FIFO, deframing and dispatching
was all done from interrupt handler. TX was done with interrupts
globally disabled, and it was far from healthy.

In this commit irq handler only posts message to os task that there
is work to be done, everything else is done from this os-level task.

Since tx can be done from different contexts and disabling interrupts
globally is not an option anymore, messages to transmit are queued
in a ringbuffer to be sent later from the task.